backend/requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
environ==1.0
idna==3.6
isodate==0.6.1
lxml==5.1.0
Markdown==3.5.2
platformdirs==4.2.0
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.2.1
zeep==4.2.1
